# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LMS Backend API",
    "version": "1.0.0",
    "description": "API documentation cho hệ thống Learning Management System (LMS)",
    "contact": {
      "name": "LMS Team",
      "email": "admin@lms.com"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.lms.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Nhập JWT token để xác thực"
      }
    },
    "schemas": {
...
      "Grade": {
        "type": "object",
        "required": [
          "studentId",
          "subjectId",
          "classId",
          "gradeValue",
          "gradeType",
          "term"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID duy nhất của điểm"
          },
          "studentId": {
            "type": "integer",
            "description": "ID học sinh"
          },
          "subjectId": {
            "type": "integer",
            "description": "ID môn học"
          },
          "classId": {
            "type": "integer",
            "description": "ID lớp học"
          },
          "gradeValue": {
            "type": "number",
            "minimum": 0,
            "maximum": 10,
            "description": "Điểm số"
          },
          "gradeType": {
            "type": "string",
            "enum": [
              "homework",
              "quiz",
              "midterm",
              "final",
              "assignment",
              "participation"
            ],
            "description": "Loại điểm"
          },
          "weight": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "description": "Trọng số"
          },
          "term": {
            "type": "string",
            "enum": [
              1,
              2,
              "final"
            ],
            "description": "Học kỳ"
          },
          "academicYear": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{4}$",
            "description": "Năm học (YYYY-YYYY)"
          },
          "remarks": {
            "type": "string",
            "maxLength": 1000,
            "description": "Ghi chú"
          },
          "recordedBy": {
            "type": "integer",
            "description": "ID người nhập điểm"
          },
          "recordedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Thời gian nhập điểm"
          },
          "isActive": {
            "type": "boolean",
            "description": "Trạng thái hoạt động"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
...
  },
...
  "paths": {
...
    "/api/v1/grades": {
      "post": {
        "tags": [
          "Grades"
        ],
        "summary": "Nhập điểm mới",
        "description": "Tạo điểm mới cho học sinh (chỉ Admin và Teacher)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "studentId",
                  "subjectId",
                  "classId",
                  "gradeValue",
                  "gradeType",
                  "term"
                ],
                "properties": {
                  "studentId": {
                    "type": "integer",
                    "description": "ID học sinh"
                  },
                  "subjectId": {
                    "type": "integer",
                    "description": "ID môn học"
                  },
                  "classId": {
                    "type": "integer",
                    "description": "ID lớp học"
                  },
                  "gradeValue": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 10,
                    "description": "Điểm số"
                  },
                  "gradeType": {
                    "type": "string",
                    "enum": [
                      "homework",
                      "quiz",
                      "midterm",
                      "final",
                      "assignment",
                      "participation"
                    ],
                    "description": "Loại điểm"
                  },
                  "weight": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 100,
                    "description": "Trọng số (tự động nếu không cung cấp)"
                  },
                  "term": {
                    "type": "string",
                    "enum": [
                      1,
                      2,
                      "final"
                    ],
                    "description": "Học kỳ"
                  },
                  "academicYear": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{4}$",
                    "description": "Năm học (tự động nếu không cung cấp)"
                  },
                  "remarks": {
                    "type": "string",
                    "maxLength": 1000,
                    "description": "Ghi chú"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Nhập điểm thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Grade"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "get": {
        "tags": [
          "Grades"
        ],
        "summary": "Lấy danh sách điểm theo lớp",
        "description": "Lấy danh sách điểm với phân trang và bộ lọc (Admin xem tất cả, Teacher chỉ xem lớp mình dạy)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "classId",
            "schema": {
              "type": "integer"
            },
            "description": "ID lớp học (optional)"
          },
          {
            "in": "query",
            "name": "subjectId",
            "schema": {
              "type": "integer"
            },
            "description": "ID môn học"
          },
          {
            "in": "query",
            "name": "studentId",
            "schema": {
              "type": "integer"
            },
            "description": "ID học sinh"
          },
          {
            "in": "query",
            "name": "term",
            "schema": {
              "type": "string",
              "enum": [
                "1",
                "2",
                "final"
              ]
            },
            "description": "Học kỳ"
          },
          {
            "in": "query",
            "name": "academicYear",
            "schema": {
              "type": "string"
            },
            "description": "Năm học (VD 2023-2024)"
          },
          {
            "in": "query",
            "name": "gradeType",
            "schema": {
              "type": "string",
              "enum": [
                "homework",
                "quiz",
                "midterm",
                "final",
                "assignment",
                "participation"
              ]
            },
            "description": "Loại điểm"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Số trang"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Số lượng kết quả mỗi trang"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "default": "recordedAt"
            },
            "description": "Sắp xếp theo trường"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ],
              "default": "DESC"
            },
            "description": "Thứ tự sắp xếp"
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách điểm thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Grade"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "currentPage": {
                              "type": "integer"
                            },
                            "totalPages": {
                              "type": "integer"
                            },
                            "totalItems": {
                              "type": "integer"
                            },
                            "itemsPerPage": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/v1/grades/report/export": {
      "get": {
        "tags": [
          "Grades"
        ],
        "summary": "Xuất báo cáo điểm Excel",
        "description": "Xuất báo cáo điểm dưới dạng file Excel",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "classId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID lớp học"
          },
          {
            "in": "query",
            "name": "subjectId",
            "schema": {
              "type": "integer"
            },
            "description": "ID môn học"
          },
          {
            "in": "query",
            "name": "term",
            "schema": {
              "type": "string",
              "enum": [
                1,
                2,
                "final"
              ]
            },
            "description": "Học kỳ"
          },
          {
            "in": "query",
            "name": "academicYear",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{4}$"
            },
            "description": "Năm học"
          }
        ],
        "responses": {
          "200": {
            "description": "File Excel báo cáo điểm",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  }
}